% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_decay.R
\name{dist_decay}
\alias{dist_decay}
\title{Distance-decay of similarity}
\usage{
dist_decay(comm, xy_coords, method = "bray", binary = F)
}
\arguments{
\item{comm}{Community matrix with samples in rows and species in columns.
The matrix can contain presence/absence data (coded with 0/1) or
abundance data with counts of individuals.}

\item{xy_coords}{Two-column table with x-y coordinates of the samples.}

\item{method}{Choice of (dis)similarity index. See \code{\link[vegan]{vegdist}}}

\item{binary}{Perform presence/absence standardization before analysis?
See \code{\link[vegan]{vegdist}}}
}
\value{
The function returns a list of two dataframes. The first, \code{dd_data} includes
all pairwise distances and their respective similarities.
The second, \code{dd_smooth} includes a fitted GAM
\code{\link[mgcv]{gam}} for the relationship between distance and
similarity.
}
\description{
Estimate pairwise similarities of communities in samples as
function of distance
}
