% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullmodel_3.R
\name{null_model}
\alias{null_model}
\title{library(vegan)
library(tidyverse)
library(ggplot2)
library(mobr)
library(spatialbiodiv)
library(scam)
setwd("C:/Users/aylin/Desktop/frag_mod_scale-main/simulation_1")
simdat0 <- read.csv("sSBR7_rep_1_output_sample.csv", header=TRUE, sep = ",")
next step: isolate the sSBR function into an additional one so it's cleaner!}
\usage{
null_model(
  dataset,
  frag_level_low,
  frag_level_high,
  timestep,
  permutations,
  sample_size
)
}
\arguments{
\item{timestep}{41 is currently directly after frag and 100 is at the end of sim
this is what will become the ribbon
when testing parts of the code, use these presets:
dataset <- simdat0
frag_level_low <- 0.1
frag_level_high <- 0.9
timestep <- 41
sample_size <- 50
this runs the sSBR for n amount of permutations
the min_dist code from Felix's Github
we don't actually need this at the moment so i commented it out

\if{html}{\out{<div class="sourceCode">}}\preformatted{ min_dist <- round(min(c(max(sSBR_test_high$sSBR_data$distance),
                         max(sSBR_test_high$sSBR_data$distance))))
}\if{html}{\out{</div>}}

instead we're using new_dist to have the same distances for all permutations
i'm creating a table that grows with each permutation so i can later use quantile()
on all permutations for distance x
TIP: in a table \link{x,y} stands for \link{row,column}
this is the actual curve that we compare to the nullmodel
initially, i created the ribbon with the min and max value of the diff_sSBR throughout
all permutations, but in order to solve Felix's request for showing the 95\% and 5\%
percentile, i now use the 95\% of all permutations as the ymax and the 5\% as the ymin
and r wants me to transform nd to a matrix instead of data.table
TIP: in a table \link{x,y} stands for \link{row,column}
this saves the distances as one column because geom_ribbon needs an x value
looping through all permutations to create the data.set used for my ribbon

I think the for loop is not necessary, because quantile can deal with vectores
this is now the plot: we have the blue rarefaction curve of the sample we want to test
against the nullmodel and the transparent green null ribbon, with additional lines for the
min and max value of the nullmodel (i don't know if these are necessary?)
printing the result!}

\item{data}{the data frame we want to look at, e.g. stav's output}

\item{frag_level}{we currently look at frag = 0.1 and 0.9, has to be in the data frame}
}
\description{
sSBR_curve = function(dataset, frag_level_low, frag_level_high, timestep,
permutations, sample_size)
the parameters of the function:
}
