% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_rarefaction_curve.R
\name{sSBR}
\alias{sSBR}
\title{Spatially-explicit sample-based species rarefaction curve}
\usage{
sSBR(comm, xy_coords)
}
\arguments{
\item{comm}{Community matrix with samples in rows and species in columns.
The matrix can contain presence/absence data (coded with 0/1) or
abundance data with counts of individuals.}

\item{xy_coords}{Two-column table with x-y coordinates of the samples.}
}
\value{
The function returns a list of two dataframes. The first, \code{sSBR_data} includes
        all accumulated distances and their respective accumulated species richness for
        the same number of rarefaction curves as there are samples in the data set.
        Each samples is used as a starting point and then samples are accumulated
        by distances to the k nearest neighbour.
        The second dataframe, \code{ssBR_smooth} includes a non-linear smoother
       \code{\link[stats]{loess}} for the relationship between cimulative distance and
       cumulative species richness.
}
\description{
This function calculates the spatial and sample-based species rarefaction curve
(sSBR after McGlinn et al. 2019), also called spatially-constrained
rarefaction curve (SCR) (Chiarucci et al. 2009).
}
\details{
In contrast to McGlinn et al. 2019, this function uses accumulated distances
among samples as x-axis and not the accumulated number of samples.
}
\references{
McGlinn, D. J. et al. 2019. Measurement of Biodiversity (MoB):
 A method to separate the scale-dependent effects of species abundance distribution,
  density, and aggregation on diversity change. - Methods in Ecology and Evolution 10: 258–269.

 Chiarucci, A. et al. 2009. Spatially constrained rarefaction:
 incorporating the autocorrelated structure of biological communities into
  sample-based rarefaction. - Community Ecology 10: 209–214.
}
